add if test for sign in and bugg fix for items in cart

+++ app/controllers/shopping_carts_controller.rb
@@ -3,7 +3,9 @@ class ShoppingCartsController < ApplicationController
 
   def complete
     @cart.update(paid: 'true')
-    session.delete(:cart_id)
+    if @cart.paid = 'true'
+      session.delete(:cart_id)
+    end
   end
 
   def show
